requirements.txt
asgiref==3.10.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
dotenv==0.9.9
Markdown==3.10
pillow==10.4.0
python-dotenv==1.2.1
sqlparse==0.5.3
